+++ tests/integration/handlers/live_collection/test_unset_operator.py
@@ -1,0 +1,9 @@
+import asynctest
+
+from tests.integration.handlers.live_collection.base import \
+    LiveCollectionFieldUpdateOperatorsTests
+
+
+class UnsetOperationTests(LiveCollectionFieldUpdateOperatorsTests,
+                          asynctest.TestCase):
+    pass
